feat: refactor server tools
- seperate server tools into their own file
